Add files via upload
Project Title: Breast Cancer Classification using Logistic Regression

This project focuses on building a logistic regression model using Scikit-learn to classify breast cancer as either malignant or benign. The Breast Cancer Wisconsin (Diagnostic) Data Set from Kaggle is used as the dataset for analysis. The goal of this project is to develop a simple logistic regression classifier that can accurately predict whether a patient has cancer or not.

The project begins with data preparation and cleaning, ensuring that the dataset is suitable for analysis. Python programming language is utilized for these tasks, leveraging its data mining and manipulation capabilities. The dataset is thoroughly examined and any missing values or inconsistencies are addressed to ensure reliable results.

Once the data is cleaned, the logistic regression algorithm is implemented using Scikit-learn, a popular machine learning library in Python. The theoretical understanding of logistic regression is applied practically to train the model on the prepared dataset. This involves splitting the data into training and testing sets, fitting the logistic regression model on the training data, and evaluating its performance on the testing data.

The model's performance is assessed using various evaluation metrics such as accuracy, precision, recall, and F1 score. These metrics provide insights into how well the model is able to classify between cancerous and non-cancerous patients. The results are interpreted and analyzed to gain a deeper understanding of the model's predictive capabilities.

Throughout the project, Python programming skills are utilized to implement the logistic regression model, preprocess the data, and perform data mining tasks. The project demonstrates My proficiency in cancer prediction (Predictive Analytics), machine learning, and data mining techniques.